upstream: b=master,r=93b64ed2ee517c14a3ad7ca38110883b38f9b46d,t=2019-08-16-0357-04143

+++ plugins/nexus-blobstore-tasks/src/main/java/org/sonatype/nexus/blobstore/restore/BaseRestoreBlobStrategy.java
@@ -113,7 +113,9 @@ public abstract class BaseRestoreBlobStrategy<T>
           log.debug(
               "Deleting asset as component is required but is not found, blob store: {}, repository: {}, path: {}, "
               + "blob name: {}, blob id: {}", blobStoreName, repoName, path, blobName, blob.getId());
-          deleteAsset(blobData.getRepository(), path);
+          if (!isDryRun) {
+            deleteAsset(blobData.getRepository(), path);
+          }
         }
         else {
           log.debug(
